pyproject.toml
[tool.poetry]
name = "speedtest_csv"
version = "0.2.0"
description = "At-home internet speed -> CSV program"
authors = ["Kevin Kredit"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.7"
speedtest-cli = "^2.1.3"
Gooey = "^1.0.8"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
pylint = "^2.9.3"
mypy = "^0.910"
pre-commit = "^2.13.0"
bandit = "^1.7.0"
isort = "^5.9.1"

[tool.black]
target-version = ["py37"]
exclude = '''
/(  \.git
  | \.mypy_cache
  | \.venv
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
